+++ src/pages/MyPage/MyPage.jsx
@@ -1,11 +1,12 @@
-import * as S from './MyPage.style';
-import Icon from '../../components/Icon/Icon';
-import React, { useState, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
+import React, { useEffect, useState } from 'react';
 import { HiOutlineLink } from 'react-icons/hi';
-import MY_CARD_SAMPLE_DATA from '../../constants/myCardSampleData';
+import { useNavigate } from 'react-router-dom';
 import { getMyCard } from '../../apis';
 import ProfileImgDefault from '../../assets/images/profile-img-default.svg';
+import Icon from '../../components/Icon/Icon';
+import MY_CARD_SAMPLE_DATA from '../../constants/myCardSampleData';
+import { formatPhoneNumber } from '../../utils/HomePageUtils/formatPhoneNumber';
+import * as S from './MyPage.style';
 
 export default function MyPage() {
   const [isModalOpen, setIsModalOpen] = useState(false);
@@ -106,7 +107,9 @@ export default function MyPage() {
               <S.MyInfoItem>
                 <S.MyInfoLabel>휴대폰</S.MyInfoLabel>
                 <S.ContactWrapper>
-                  <S.MyInfoValue>{myInfo.phone}</S.MyInfoValue>
+                  <S.MyInfoValue>
+                    {formatPhoneNumber(myInfo.phone)}
+                  </S.MyInfoValue>
                 </S.ContactWrapper>
               </S.MyInfoItem>
               {myInfo.email && (
@@ -119,7 +122,9 @@ export default function MyPage() {
                 <S.MyInfoItem>
                   <S.MyInfoLabel>유선전화</S.MyInfoLabel>
                   <S.ContactWrapper>
-                    <S.MyInfoValue>{myInfo.tel}</S.MyInfoValue>
+                    <S.MyInfoValue>
+                      {formatPhoneNumber(myInfo.tel)}
+                    </S.MyInfoValue>
                   </S.ContactWrapper>
                 </S.MyInfoItem>
               )}
